MBS-11392 merge TestExecutor and TestsRunner into TestScheduler
