+++ subprojects/test-runner/client/src/main/kotlin/com/avito/runner/config/InstrumentationTestsActionParams.kt
@@ -3,7 +3,7 @@ package com.avito.runner.config
 import com.avito.android.runner.report.ReportViewerConfig
 import com.avito.android.stats.StatsDConfig
 import com.avito.logger.LoggerFactory
-import com.avito.runner.scheduler.runner.ExecutionParameters
+import com.avito.runner.scheduler.runner.model.ExecutionParameters
 import com.avito.runner.scheduler.suite.filter.ImpactAnalysisResult
 import com.avito.utils.gradle.KubernetesCredentials
 import java.io.File
