--- subprojects/test-runner/client/src/main/kotlin/com/avito/runner/scheduler/runner/TestExecutor.kt
@@ -1,17 +1,0 @@
-package com.avito.runner.scheduler.runner
-
-import com.avito.runner.scheduler.runner.model.TestWithTarget
-import java.io.File
-
-public interface TestExecutor {
-
-    public fun execute(
-        application: File?,
-        testApplication: File,
-        testsToRun: List<TestWithTarget>,
-        executionParameters: ExecutionParameters,
-        output: File
-    )
-
-    public companion object
-}
