+++ subprojects/test-runner/client/src/main/kotlin/com/avito/runner/scheduler/runner/model/TargetGroup.kt
@@ -1,0 +1,5 @@
+package com.avito.runner.scheduler.runner.model
+
+import com.avito.runner.config.Reservation
+
+internal data class TargetGroup(val name: String, val reservation: Reservation)
