--- subprojects/test-runner/client/src/main/kotlin/com/avito/runner/scheduler/runner/scheduler/TestsRunner.kt
@@ -1,15 +1,0 @@
-package com.avito.runner.scheduler.runner.scheduler
-
-import com.avito.android.runner.report.Report
-import com.avito.runner.scheduler.runner.model.TestWithTarget
-import java.io.File
-
-internal interface TestsRunner {
-
-    fun runTests(
-        mainApk: File?,
-        testApk: File,
-        report: Report,
-        testsToRun: List<TestWithTarget>
-    )
-}
