+++ subprojects/test-runner/client/src/testFixtures/kotlin/com/avito/runner/scheduler/runner/StubRunnerParameters.kt
@@ -1,5 +1,7 @@
 package com.avito.runner.scheduler.runner
 
+import com.avito.runner.scheduler.runner.model.ExecutionParameters
+
 public fun ExecutionParameters.Companion.createStubInstance(
     applicationPackageName: String = "com.test.app",
     applicationTestPackageName: String = "com.test.app.test",
